[FIXED] Keep the sorting of non priority Slots
We should now be back to normal AND have the
possibility to add @signalPriority to Slots.

Sadly we can NOT use usort for this kind of
sorting because Zends Bubblesort is always
unstable and we can't predict the sorting
for equal priorities.

See: http://php.net/manual/en/function.usort.php

Fixes #57